* Fixed: Compiler warning.
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@6080 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/tools/ConsoleRunner/main.cpp
@@ -88,7 +88,7 @@ int main(int argc, char** argv)
     }
 
     // count size of arguments
-    int fullsize = 0;
+    size_t fullsize = 0;
     for (int i = 1; i < argc; ++i)
     {
         fullsize += strlen(argv[i]);
